requirements.txt
bcrypt==4.2.1
blinker==1.8.2
click==8.1.8
flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
flask-sqlalchemy==3.1.1
greenlet==3.1.1
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.5
MarkupSafe==2.1.5
mysqlclient==2.2.7
python-dotenv==1.0.1
SQLAlchemy==2.0.37
typing-extensions==4.12.2
werkzeug==3.0.6
zipp==3.20.2
